Add note about eagerly detecting disconnections
This could be done by adding a separate goroutine that continually reads
from the connection. If the connection is severed or closed, the read
request will fail more quickly and break the client loose. Otherwise the
client will have to wait for socket- or application-level keepalives to
notice the broken connection.

For now, we can rely on the ping/pong requests used by the sauron client.