Flesh out cDNS service
This relies on containers having metadata labels indicating which tenant
they belong to and which services they expose, which means the service
requires Docker 1.6 or later.

It would also be possible to configure this data via environment variables,
but using labels means we don't have to expose this information inside the
container, and we're going to have to upgrade our Docker version at some
point.