Resurrect the CLI admin interface
Unlike the previous implementation, the client now has a canned set of
commands it knows about, and all communication is done via the REST-ish
interface of the server's domain socket.

This is roughly similar to how the Docker remote API works, except this CLI
binary uses a more traditional REPL interface.