Use more lenient path matching in admin REST API
This fixes an escaping issue in the client used by the CLI when peer or
connection names also happen to contain path characters. Clients can work
around this by ensuring these characters are encoded before they are sent,
but it is more convenient to allow "/peers/foo%2fbar" to be sent as
"/peers/foo/bar".