Fix removing peers through the CLI
The default behavior for ServerMux, inherited when using Gorilla's router,
is to "clean" submitted URLs by returning a 301 when clients request a URL
that happens to contain redundant path information. For most cases this is
fine, but when peer URLs are encoded in paths, this behavior mangles the
submitted address making it impossible to for the client to convey an
address such as "tcp://host.address:port".

Rather than work around this upstream behavior, this commit makes the
client API less RESTful by adding explicit POST routes for adding and
deleting peers. For consistency this commit also modifies the connection
routes similarly.